pyproject.toml
[tool.poetry]
name = "doom"
version = "0.1.0"
description = ""
authors = ["Bill Tyros <me@strategineer.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
omgifol = "^0.5.1"
pillow = "^11.0.0"
pyfluidsynth = "^1.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
